abcxyz

+++ Login/Performance/Store_performence.cs
@@ -48,7 +48,7 @@ namespace Login
                 string Ten = reader.GetString(1).Trim();
                 double gia = reader.GetFloat(2);
                 int soluong = reader.GetInt32(3);
-                int danhgia = reader.GetInt32(4);
+                double danhgia = reader.GetFloat(4);
                 int daban = reader.GetInt32(5);
                 string mota = reader.GetString(6);
                 string Loai = reader.GetString(7);
